Extract CityMap request handling module
